:hammer: chore: move getCADVersion to utils
