+++ packages/api/src/controllers/admin/manage/cad-settings/CadSettings.ts
@@ -15,7 +15,7 @@ import { Socket } from "services/SocketService";
 import { nanoid } from "nanoid";
 import { validateSchema } from "lib/validateSchema";
 import type { cad, Feature } from "@prisma/client";
-import { getCADVersion } from "src/main";
+import { getCADVersion } from "@snailycad/utils/version";
 
 @Controller("/admin/manage/cad-settings")
 export class ManageCitizensController {
