+++ packages/api/src/main.ts
@@ -1,9 +1,7 @@
 import { $log } from "@tsed/logger";
 import { PlatformExpress } from "@tsed/platform-express";
 import { Server } from "./server";
-import { resolve } from "node:path";
-import { readFile } from "node:fs/promises";
-import process from "node:process";
+import { getCADVersion } from "@snailycad/utils/version";
 
 async function bootstrap() {
   try {
@@ -17,19 +15,6 @@ async function bootstrap() {
   } catch (er) {
     $log.error(er);
   }
-}
-
-let versionCache: string;
-export async function getCADVersion(): Promise<string | null> {
-  const packageJsonPath = resolve(process.cwd(), "package.json");
-  const packageJson = await readFile(packageJsonPath, "utf-8").catch(() => null);
-  if (!packageJson) return null;
-
-  const json = JSON.parse(packageJson);
-
-  versionCache ??= json.version;
-
-  return versionCache;
 }
 
 bootstrap();
