+++ packages/telemetry/package.json
@@ -6,6 +6,9 @@
     "build": "tsup",
     "typecheck": "tsc --noEmit"
   },
+  "dependencies": {
+    "@snailycad/utils": "1.0.0-beta.77"
+  },
   "devDependencies": {
     "@types/node": "^17.0.23",
     "tslib": "^2.3.1",
