+++ packages/telemetry/src/index.ts
@@ -2,8 +2,10 @@ import process from "node:process";
 import os from "node:os";
 import { execSync } from "node:child_process";
 import { get, set } from "./cache";
+import { getCADVersion } from "@snailycad/utils/version";
 
 const TELEMETRY_ENABLED = process.env.TELEMETRY_ENABLED === "true";
+const REPORT_URL = "";
 
 interface ErrorReport {
   name: string;
@@ -14,22 +16,30 @@ interface ErrorReport {
 export async function report(errorReport: ErrorReport) {
   if (!TELEMETRY_ENABLED) return;
 
-  const [yarn, node, npm] = await Promise.all([
+  const [yarn, node, npm, cadVersion] = await Promise.all([
     getBinaryVersions("yarn"),
     getBinaryVersions("node"),
     getBinaryVersions("npm"),
+    getCADVersion(),
   ]);
 
   const data = {
     yarn,
     npm,
     node,
+    cadVersion,
     platform: os.platform(),
     os: os.platform(),
     ...errorReport,
   };
 
-  data;
+  const res = await fetch(REPORT_URL, {
+    method: "POST",
+    body: JSON.stringify(data),
+    cache: "no-cache",
+  });
+
+  console.log({ res });
 }
 
 async function getBinaryVersions(command: "yarn" | "node" | "npm") {
