+++ packages/utils/tests/version.test.ts
@@ -1,0 +1,7 @@
+/* eslint-disable quotes */
+import { expect, test } from "vitest";
+import { getCADVersion } from "../src/version";
+
+test("Should return the CAD version", async () => {
+  expect(await getCADVersion()).toBeTypeOf("string");
+});
