+++ packages/utils/tsconfig.json
@@ -7,6 +7,7 @@
     "moduleResolution": "node",
     "declaration": true,
     "skipLibCheck": true,
-    "importHelpers": true
+    "importHelpers": true,
+    "esModuleInterop": true
   }
 }
